% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minstrel-package.R
\docType{package}
\name{minstrel-package}
\alias{minstrel}
\alias{minstrel-package}
\title{minstrel: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Brady Johnston \email{brady.johnston@me.com} (\href{https://orcid.org/0000-0001-6301-2269}{ORCID})

}
\keyword{internal}
